Projects/IMDB_Spoiler_Shield/requirements.txt
pandas
numpy
scikit-learn
torch
transformers
datasets
mlflow==2.14.1
apache-airflow==2.10.0
evidently
nltk
psycopg2-binary
boto3
gdown
matplotlib
seaborn
wordcloud
tqdm
feast
redis
s3fs
xgboost
sentence-transformers
ray[default]
ray[tune]
ray[train]
accelerate
